requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
django_filter==23.2
django-rest-auth[with_social]==0.9.5
django-dotenv==1.4.2
pillow==10.0.0
